feat: add should_email field to notifications
