+++ src/services/broadcastCoordinationService.server.ts
@@ -1,5 +1,5 @@
 import { discordServiceServer } from './discordService.server'
-import { notificationsSupabaseServiceServer } from './notificationSupabaseService.server'
+import { notificationsSupabaseServiceServer } from './notificationsSupabaseService.server'
 
 import type { SupabaseClient } from '@supabase/supabase-js'
 import type { DBProfile, DBResearchUpdate, ResearchUpdate } from 'oa-shared'
