+++ src/test/services/notificationEmailService.server.test.ts
@@ -52,29 +52,6 @@ describe('notificationEmailService', () => {
       expect(mocks.functionsInvoke).not.toHaveBeenCalled()
     })
 
-    it('should return early when shouldSendEmail returns false', async () => {
-      const { client, mocks } = createMockSupabaseClient()
-      const notification = factoryDBNotification()
-
-      mocks.single.mockResolvedValue({
-        data: { created_at: '2024-01-01', roles: ['beta-tester'] },
-      })
-
-      mockGetPreferences.mockResolvedValue({ is_unsubscribed: true })
-
-      const result =
-        await notificationEmailService.createInstantNotificationEmail(
-          client,
-          notification,
-          123,
-          headers,
-        )
-
-      expect(result).toBeUndefined()
-      expect(mocks.rpc).not.toHaveBeenCalled()
-      expect(mocks.functionsInvoke).not.toHaveBeenCalled()
-    })
-
     it('should error when RPC response has no data', async () => {
       const { client, mocks } = createMockSupabaseClient()
       const notification = factoryDBNotification()
@@ -192,204 +169,6 @@ describe('notificationEmailService', () => {
           },
         },
       })
-    })
-  })
-
-  describe('shouldSendEmail behaviour', () => {
-    it('should use default preferences when user preferences not found', async () => {
-      const { client, mocks } = createMockSupabaseClient()
-      const notification = factoryDBNotification()
-
-      mocks.single.mockResolvedValue({
-        data: { created_at: '2024-01-01', roles: ['beta-tester'] },
-      })
-
-      mockGetPreferences.mockResolvedValue(null)
-      mocks.rpc.mockResolvedValue({
-        data: [{ email: 'test@example.com' }],
-      })
-
-      mockTransformNotification.mockResolvedValue(notification)
-      mockTokensGenerate.mockReturnValue('token')
-
-      await notificationEmailService.createInstantNotificationEmail(
-        client,
-        notification,
-        123,
-        headers,
-      )
-
-      expect(mocks.functionsInvoke).toHaveBeenCalled()
-    })
-
-    it('should return false when user is unsubscribed', async () => {
-      const { client, mocks } = createMockSupabaseClient()
-      const notification = factoryDBNotification()
-
-      mocks.single.mockResolvedValue({
-        data: { created_at: '2024-01-01', roles: ['beta-tester'] },
-      })
-
-      mockGetPreferences.mockResolvedValue({
-        is_unsubscribed: true,
-        comments: true,
-      })
-
-      const result =
-        await notificationEmailService.createInstantNotificationEmail(
-          client,
-          notification,
-          123,
-          headers,
-        )
-
-      expect(result).toBeUndefined()
-      expect(mocks.rpc).not.toHaveBeenCalled()
-      expect(mocks.functionsInvoke).not.toHaveBeenCalled()
-    })
-
-    it('should use user preference when available and not unsubscribed', async () => {
-      const { client, mocks } = createMockSupabaseClient()
-      const notification = factoryDBNotification()
-
-      mocks.single.mockResolvedValue({
-        data: { created_at: '2024-01-01', roles: ['beta-tester'] },
-      })
-
-      mockGetPreferences.mockResolvedValue({
-        is_unsubscribed: false,
-        comments: true,
-      })
-      mocks.rpc.mockResolvedValue({
-        data: [{ email: 'test@example.com' }],
-      })
-
-      mockTransformNotification.mockResolvedValue(notification)
-      mockTokensGenerate.mockReturnValue('token')
-
-      await notificationEmailService.createInstantNotificationEmail(
-        client,
-        notification,
-        123,
-        headers,
-      )
-
-      expect(mocks.functionsInvoke).toHaveBeenCalled()
-    })
-
-    it('should not send email when user preference is false', async () => {
-      const { client, mocks } = createMockSupabaseClient()
-      const notification = factoryDBNotification()
-
-      mocks.single.mockResolvedValue({
-        data: { created_at: '2024-01-01', roles: ['beta-tester'] },
-      })
-
-      mockGetPreferences.mockResolvedValue({
-        is_unsubscribed: false,
-        comments: false,
-      })
-
-      const result =
-        await notificationEmailService.createInstantNotificationEmail(
-          client,
-          notification,
-          123,
-          headers,
-        )
-
-      expect(result).toBeUndefined()
-      expect(mocks.rpc).not.toHaveBeenCalled()
-      expect(mocks.functionsInvoke).not.toHaveBeenCalled()
-    })
-
-    it('should handle reply content type correctly', async () => {
-      const { client, mocks } = createMockSupabaseClient()
-      const notification = factoryDBNotification({ content_type: 'reply' })
-
-      mocks.single.mockResolvedValue({
-        data: { created_at: '2024-01-01', roles: ['beta-tester'] },
-      })
-
-      mockGetPreferences.mockResolvedValue({
-        is_unsubscribed: false,
-        replies: true,
-      })
-      mocks.rpc.mockResolvedValue({
-        data: [{ email: 'test@example.com' }],
-      })
-
-      mockTransformNotification.mockResolvedValue(notification)
-      mockTokensGenerate.mockReturnValue('token')
-
-      await notificationEmailService.createInstantNotificationEmail(
-        client,
-        notification,
-        123,
-        headers,
-      )
-
-      expect(mocks.functionsInvoke).toHaveBeenCalled()
-    })
-
-    it('should handle researchUpdate content type correctly', async () => {
-      const { client, mocks } = createMockSupabaseClient()
-      const notification = factoryDBNotification({
-        content_type: 'researchUpdate',
-      })
-
-      mocks.single.mockResolvedValue({
-        data: { created_at: '2024-01-01', roles: ['beta-tester'] },
-      })
-
-      mockGetPreferences.mockResolvedValue({
-        is_unsubscribed: false,
-        research_updates: true,
-      })
-      mocks.rpc.mockResolvedValue({
-        data: [{ email: 'test@example.com' }],
-      })
-
-      mockTransformNotification.mockResolvedValue(notification)
-      mockTokensGenerate.mockReturnValue('token')
-
-      await notificationEmailService.createInstantNotificationEmail(
-        client,
-        notification,
-        123,
-        headers,
-      )
-
-      expect(mocks.functionsInvoke).toHaveBeenCalled()
-    })
-
-    it('should use default preference when user preference is undefined', async () => {
-      const { client, mocks } = createMockSupabaseClient()
-      const notification = factoryDBNotification()
-
-      mocks.single.mockResolvedValue({
-        data: { created_at: '2024-01-01', roles: ['beta-tester'] },
-      })
-
-      mockGetPreferences.mockResolvedValue({
-        is_unsubscribed: false,
-        comments: undefined,
-      })
-      mocks.rpc.mockResolvedValue({
-        data: [{ email: 'test@example.com' }],
-      })
-
-      mockTransformNotification.mockResolvedValue(notification)
-      mockTokensGenerate.mockReturnValue('token')
-
-      await notificationEmailService.createInstantNotificationEmail(
-        client,
-        notification,
-        123,
-        headers,
-      )
-
-      expect(mocks.functionsInvoke).toHaveBeenCalled()
     })
   })
 })
